ported to cdh5.10, added timeout for tool processor
The timout is used to kill a tool execution porcess after a time frame defined in the Processor class. This is required to avoid that a mapreduce job fails, if a tool hangs. The map process is otherwise killed by the system after a timeout defined by the mapred.task.timeout property (mapred-site.xml), causing the entire job to fail.